+++ packages/cron/src/jobs/pins.js
@@ -101,9 +101,8 @@ export async function updatePinStatuses ({ cluster, db }) {
       return {
         id: pin._id,
         status: status,
-        content_cid: pin.contentCid,
-        pin_location_id: pin.location._id,
-        updated_at: new Date().toISOString()
+        cid: pin.contentCid,
+        locationId: pin.location._id
       }
     }))